Use webpack for the build process
