+++ README.md
@@ -1,27 +1,40 @@
 # PWAPlanet
+
 A Progressive Web Application to load in cards posts from
 [Planet KDE](planet.kde.org).
 
+## How to 
+
+* `npm install`
+* `npm run dev` for development
+* `npm run build` for production
+	* Built files will be found in the `dist/` folder
+
 ## Version
+
 0.2
 
 ### TODO
-[x] Share content
-[ ] Bookmark article - Save the article to read later - Maybe using AMP?
-[ ] Save favorite cards - Maybe this one is better
+
+* [x] Share content
+* [ ] Bookmark article - Save the article to read later - Maybe using AMP?
+* [ ] Save favorite cards - Maybe this one is better
 
 ### Bugs
+
 [#1 - Icons Not Loaded](https://github.com/lays147/PWAPlanet/issues/1)
 [#2 - Cache don't update](https://github.com/lays147/PWAPlanet/issues/2) 
-
 ### License
+
 Project under MIT License
 
 ### Inspiration
+
 This application was inspired after I went to the PWA Roadshow made by
 Google in Brazil.
 
 ### More info About PWA
-- [Google PWA](https://developers.google.com/web/progressive-web-apps/)
-- [Google Codelabs](codelabs.developers.google.com)
-- [Google Showcases](https://developers.google.com/web/showcase/)
+
+* [Google PWA](https://developers.google.com/web/progressive-web-apps/)
+* [Google Codelabs](codelabs.developers.google.com)
+* [Google Showcases](https://developers.google.com/web/showcase/)
