+++ src/index.template.html
@@ -25,13 +25,6 @@
   <meta name="apple-mobile-web-app-title" content="Planet KDE">
   <link rel="apple-touch-icon" href="images/touch/apple-touch-icon.png">
   <meta name="theme-color" content="#2196F3">
-
-  <!-- Material Design icons -->
-  <link href="styles/icons/materialdesignicons/css/materialdesignicons.min.css" media="all" rel="stylesheet" type="text/css" />
-  <link href="styles/icons/iconfontfromgoogle/material-icons.css" media="all" rel="stylesheet" type="text/css" />
-  <!-- Your styles -->
-  <link rel="stylesheet" href="styles/material.min.css">
-  <link rel="stylesheet" href="styles/icons.css">
 </head>
 
 <body>
@@ -112,9 +105,6 @@
     </main>
     <div class="loader mdl-spinner mdl-js-spinner"></div>
   </div>
-  <script src="scripts/material.min.js"></script>
-  <!-- build:js scripts/main.min.js -->
-  <script src="scripts/main.js"></script>
   <script>
     (function(i, s, o, g, r, a, m) {
       i['GoogleAnalyticsObject'] = r;
