+++ static/humans.txt
