--- styles/icons/materialdesignicons/bower.json
@@ -1,23 +1,0 @@
-{
-    "name": "mdi",
-    "version": "1.9.33",
-    "main": [
-        "scss/materialdesignicons.scss"
-    ],
-    "homepage": "http://materialdesignicons.com",
-    "authors": [
-        { "name": "Austin Andrews", "homepage": "http://templarian.com" },
-        { "name": "Google", "homepage": "http://www.google.com/design" }
-    ],
-    "license": ["OFL-1.1", "MIT"],
-    "ignore": [
-        "*.md",
-        "*.json"
-    ],
-    "keywords": [
-        "material",
-        "design",
-        "icons",
-        "webfont"
-    ]
-}
