--- styles/icons/materialdesignicons/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "mdi",
-  "version": "1.9.33",
-  "description": "Dist for Material Design Webfont. This includes the Stock and Community icons in a single webfont collection.",
-  "main": "preview.html",
-  "style": "css/materialdesignicons.css",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Templarian/MaterialDesign-Webfont.git"
-  },
-  "keywords": [
-    "material",
-    "design",
-    "icons",
-    "webfont"
-  ],
-  "author": {
-    "name": "Austin Andrews",
-    "web": "http://twitter.com/templarian"
-  },
-  "licenses": [
-    {
-      "type": "OFL-1.1",
-      "url": "http://scripts.sil.org/OFL"
-    },
-    {
-      "type": "MIT",
-      "url": "http://opensource.org/licenses/mit-license.html"
-    }
-  ],
-  "bugs": {
-    "url": "https://github.com/Templarian/MaterialDesign/issues"
-  },
-  "homepage": "http://materialdesignicons.com"
-}
