--- styles/icons/materialdesignicons/scss/_icons.scss
@@ -1,5 +1,0 @@
-@each $key, $value in $mdi-icons {
-    .#{$mdi-css-prefix}-#{$key}:before {
-        content: char($value);
-    }
-}
