--- styles/icons/materialdesignicons/scss/materialdesignicons.scss
@@ -1,8 +1,0 @@
-/* MaterialDesignIcons.com */
-@import "variables";
-@import "functions";
-@import "path";
-@import "core";
-@import "icons";
-@import "extras";
-@import "animated";
